agent: Move network device code to network_device_handler
Move network device code to network_device_handler to simplify the code.
